environの生成をクラスへ
